refactor: way to retrieve pull request title
Note: Previous implementation required one to specify that
a workflow had to be triggered on pull request edit if one
wanted to trigger the changelog action after editing a pull
request title. However, this comes with triggering the whole
workflow, which can be composed of multiple time consuming jobs.
This approach allows one to rerun only the changelog action
after title changes.