Make the checkout step optional
It can be useful to skip the checkout if there is a prior step that does some setup. For example, a project might have a cython step that compiles some modules which should be packaged into the wheelhouse. This can be done in a prior step and then this action triggered with the checkout disabled.